requirements.txt
Django==1.9.7
django-factory-boy==1.0.0
factory-boy==2.7.0
fake-factory==0.5.7
psycopg2==2.6.1
python-dateutil==2.5.3
six==1.10.0
